requirements.txt
openai==1.6.1
pypdf==4.1.0
qdrant-client
langchain==0.1.11
literalai==0.0.204
chainlit==1.0.300
python-dotenv==1.0.0
tiktoken==0.5.2
langchain-openai
